+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Read commandline arguments from files:  You can now specify a file using
+  `@filename` and each line from that file will be interpreted as a commandline
+  argument.
 
 
 ## [0.6.6] - 2019-01-18