+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 - Read commandline arguments from files:  You can now specify a file using
   `@filename` and each line from that file will be interpreted as a commandline
   argument.
-- A [man-page](./doc/tbot.1)!
+- A man-page: `doc/tbot.1`!
+
+### Fixed
+- `boot_to_shell` is no longer a public method of `BoardLinux` machines.
 
 
 ## [0.6.6] - 2019-01-18